Initial upload of DITADTD folder
This is the standard DITA 1.1 set augmented by the IBM specialization
examples from the original SourceForge project and by some
specializations used in demo expeDITA applications. In addition, this
package includes the Oct 7 2015 version of Lightweight DITA. All these
items are represented in the catalog-dita.xml file. Its installed path
must be cited in the XML_CATALOG_FILES environment variable declared in
the OS-specific manner.

Little change is expected in the core parts of this set, but updates for
new specializations and for the lw-dita plugin are of course very likely
in the near term.